+++ RELEASE_NOTES.md
@@ -2,7 +2,8 @@
 
 ## v1.0.2
 - Enabled Controller security group management
-  
+- Fix issue where NSG throws an error getting a null in stead of a string, when disabling deployment of copilot.
+
 ## v1.0.1
 - Remove unnecessary data sources
 - Update copilot init module version