+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # module name - release notes
 
+## v1.1.0
+- Now supports g4 images (v8.0.0)
+
 ## v1.0.2
 - Enabled Controller security group management
 - Fix issue where NSG throws an error getting a null in stead of a string, when disabling deployment of copilot.